+++ src/CoordinateSystem12Well.php
@@ -1,0 +1,16 @@
+<?php declare(strict_types=1);
+
+namespace Mll\Microplate;
+
+class CoordinateSystem12Well extends CoordinateSystem
+{
+    public function rowCoordinates(): array
+    {
+        return range('A', 'C');
+    }
+
+    public function columnCoordinates(): array
+    {
+        return range(1, 4);
+    }
+}